Implemented SQl sever Customer Repository, changed data type of id;

+++ CustomerTransaction.SharedEntities/Models/Customer.cs
@@ -5,7 +5,7 @@ namespace CustomerTransaction.SharedEntities.Models
 {
     public class Customer
     {
-        public Int32 CustomerId { get; set; }
+        public Int64 CustomerId { get; set; }
         public String Name { get; set; }
         public String Email { get; set; }
         public Int32 Mobile { get; set; }
